Move local volume completely.
